add modules.js
put additional functions in module
