+++ miamoremustachio/test_request/modules.js
@@ -20,24 +20,8 @@ function showList(taskObject) {
     };
 }
 
-function getToDoList(amountOfTasks) {
-    try {
-        if (!isNumberValid(amountOfTasks)) {
-            throw new Error('Error: task number is invalid.');
-        };
-
-        const promises = [];
-
-        for (let i = 1; i <= amountOfTasks; i++) {
-            promises.push(requestTask(i));
-        };
-
-        Promise.all(promises)
-            .then(values => { showList(values) });
-
-    } catch(error) {
-        console.error(error.message);
-    };
-}
-
-getToDoList(10);
+module.exports = {
+    isNumberValid,
+    requestTask,
+    showList,
+};
